BCSystems/requirements.txt
dj-database-url==0.5.0
Django==3.0
django-ckeditor==6.2.0
django-contrib-comments==2.1.0
django-cors-headers==3.6.0
django-easy-pdf==0.1.1
django-extensions==3.1.5
django-graphql-jwt==0.3.4
django-js-asset==2.0.0
django-liststyle==0.1
django-mssql-backend==2.8.1
django-phone-field==1.8.1
django-pyodbc==1.1.1
django-session-timeout==0.1.0
django-sslserver==0.22
django-summernote==0.8.20.0
django_heroku
gunicorn
psycopg2
pytz
whitenoise
django-contrib-comments
Pillow-10.4.0
django-session-timeout
django-storages
django-background-tasks
pygooglevoice
django-phone-field
requests
django-user-visit==0.4
django-allauth
opencv-python
pyzbar==0.1.8
python-barcode
python-barcode[images]
cloudinary
django-cloudinary-storage
numpy-2.0.1
serial
ics
requests==2.25.1
icalendar-5.0.13
html2text-2024.2.26
openpyxl-3.1.5
